Extract validation to own file
