+++ src/react/hooks/useLoadableQuery.ts
@@ -38,6 +38,7 @@ import { __DEV__ } from "@apollo/client/utilities/environment";
 import { invariant } from "@apollo/client/utilities/invariant";
 
 import { __use, useDeepMemo, useRenderGuard } from "./internal/index.js";
+import { validateSuspenseHookOptions } from "./internal/validateSuspenseHookOptions.js";
 import { useApolloClient } from "./useApolloClient.js";
 
 type ResetFunction = () => void;
@@ -325,43 +326,6 @@ export function useLoadableQuery<
   return [loadQuery, queryRef, { fetchMore, refetch, reset, subscribeToMore }];
 }
 
-function validateOptions<TData, TVariables extends OperationVariables>(
-  options: WatchQueryOptions<TVariables, TData>
-) {
-  const { fetchPolicy, returnPartialData } = options;
-
-  validateFetchPolicy(fetchPolicy);
-  validatePartialDataReturn(fetchPolicy, returnPartialData);
-}
-
-function validateFetchPolicy(
-  fetchPolicy: WatchQueryFetchPolicy = "cache-first"
-) {
-  const supportedFetchPolicies: WatchQueryFetchPolicy[] = [
-    "cache-first",
-    "network-only",
-    "no-cache",
-    "cache-and-network",
-  ];
-
-  invariant(
-    supportedFetchPolicies.includes(fetchPolicy),
-    `The fetch policy \`%s\` is not supported with suspense.`,
-    fetchPolicy
-  );
-}
-
-function validatePartialDataReturn(
-  fetchPolicy: WatchQueryFetchPolicy | undefined,
-  returnPartialData: boolean | undefined
-) {
-  if (fetchPolicy === "no-cache" && returnPartialData) {
-    invariant.warn(
-      "Using `returnPartialData` with a `no-cache` fetch policy has no effect. To read partial data from the cache, consider using an alternate fetch policy."
-    );
-  }
-}
-
 function useWatchQueryOptions<TData, TVariables extends OperationVariables>({
   client,
   query,
@@ -386,7 +350,7 @@ function useWatchQueryOptions<TData, TVariables extends OperationVariables>({
     };
 
     if (__DEV__) {
-      validateOptions(watchQueryOptions as any);
+      validateSuspenseHookOptions(watchQueryOptions as any);
     }
 
     return watchQueryOptions as WatchQueryOptions<TVariables, TData>;
